testing auto-assign

+++ database.sql
@@ -38,7 +38,7 @@ CREATE TABLE EmployeeShiftInformation (
 	EmployeeID int NOT NULL,
     Day varchar(50) NOT NULL,
     ShiftPref varchar(50) NOT NULL,
-    NoOfHrsWorked int NOT NULL,
+    NoOfHrsWorked DOUBLE NOT NULL,
     CONSTRAINT FK_EmployeeID2 FOREIGN KEY (EmployeeID)
     REFERENCES userAccount(EmployeeID)
     on update cascade
